+++ deno-fs/mod.ts
@@ -10,6 +10,7 @@ import { copy } from "@std/fs/copy";
 import { move } from "@std/fs/move";
 import type {
   CompleteStorageModule,
+  ListItemsOptions,
   StorageKey,
   StorageModule,
 } from "@storage/common/types";
@@ -109,7 +110,7 @@ export async function removeItem(key: StorageKey): Promise<void> {
  */
 export async function* listItems<T>(
   keyPrefix: StorageKey = [],
-  _reverse = false,
+  _options?: ListItemsOptions,
 ): AsyncIterable<[StorageKey, T]> {
   const root = dirpath();
   const path = dirpath(keyPrefix);