+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0]
+
+### Changed
+
+- BREAKING: `listItems` takes a `ListItemsOptions` as it's second arg instead of
+  just `reverse` boolean
+
 ## [0.8.0]
 
 ### Changed